src/requirements.txt
functions-framework==3.*
google-cloud-pubsub
pandas
gcsfs